requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
hyperliquid-python-sdk==0.22.0
python-telegram-bot==21.9
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
